+++ src/main/java/Application.java
@@ -1,7 +1,10 @@
-package racingcar;
+import race.Run;
 
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
+
+        Run run = new Run();
+
     }
 }
