+++ src/test/java/view/ApplicationTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package view;
 
 import camp.nextstep.edu.missionutils.test.NsTest;
 import org.junit.jupiter.api.Test;
